+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## UNRELEASE
+
+- **Fix** : (OneClick) Prevent saving payment card details without proper authorization.
+
 ## 2.7.1
 
 - **Fix** : Improve the API calls to "availablePaymentProduct" to enhance performance