+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
-# UNRELEASE
+## UNRELEASE
 
+- **Fix** : (OneClick) Prevent saving payment card details without proper authorization.
 - **Fix** : Remove SDK js init from email template Multibanco/Sisal
 
 ## 2.7.1